justgram node add mvc pattern all refactoring & server.js app.js split finished
