+++ models/loginUserDao.js
@@ -1,0 +1,24 @@
+const { DataSource } = require("typeorm");
+
+const myDataSource = new DataSource({
+  type: process.env.TYPEORM_CONNECTION,
+  host: process.env.TYPEORM_HOST,
+  port: process.env.TYPEORM_PORT,
+  username: process.env.TYPEORM_USERNAME,
+  password: process.env.TYPEORM_PASSWORD,
+  database: process.env.TYPEORM_DATABASE,
+});
+
+myDataSource.initialize().then(() => {
+  console.log("Data Source has been initialized!");
+});
+
+const loginUser = async (email) => {
+  const [userInfo] = await myDataSource.query(`
+    SELECT * FROM users 
+    WHERE email ="${email}"`);
+
+  return userInfo;
+};
+
+module.exports = { loginUser };
