+++ models/postListDao.js
@@ -1,0 +1,29 @@
+const { DataSource } = require("typeorm");
+
+const myDataSource = new DataSource({
+  type: process.env.TYPEORM_CONNECTION,
+  host: process.env.TYPEORM_HOST,
+  port: process.env.TYPEORM_PORT,
+  username: process.env.TYPEORM_USERNAME,
+  password: process.env.TYPEORM_PASSWORD,
+  database: process.env.TYPEORM_DATABASE,
+});
+
+myDataSource.initialize().then(() => {
+  console.log("Data Source has been initialized!");
+});
+
+const postList = async () => {
+  const listInfo = await myDataSource.query(`SELECT
+  users.id as userId,
+  users.profile_image as userProfileImage,
+  postings.id as postingId,
+  posting_images.image_url userProfileUrl,
+  postings.contents as postingContent
+  FROM users, postings, posting_images
+  WHERE users.id = postings.user_id && postings.id = posting_images.posting_id`);
+
+  return listInfo;
+};
+
+module.exports = { postList };
