+++ models/userPostDao.js
@@ -1,0 +1,29 @@
+const { DataSource } = require("typeorm");
+
+const myDataSource = new DataSource({
+  type: process.env.TYPEORM_CONNECTION,
+  host: process.env.TYPEORM_HOST,
+  port: process.env.TYPEORM_PORT,
+  username: process.env.TYPEORM_USERNAME,
+  password: process.env.TYPEORM_PASSWORD,
+  database: process.env.TYPEORM_DATABASE,
+});
+
+myDataSource.initialize().then(() => {
+  console.log("Data Source has been initialized!");
+});
+
+const userPost = async (id) => {
+  const listInfo = await myDataSource.query(`SELECT
+  postings.user_id,
+  users.profile_image,
+  posting_images.posting_id,
+  posting_images.image_url,
+  postings.contents
+  FROM users, posting_images
+  INNER JOIN postings ON ${id} = postings.user_id`);
+
+  return listInfo;
+};
+
+module.exports = { userPost };
