+++ services/createUserService.js
@@ -41,5 +41,5 @@ const createUser = async (
 };
 
 module.exports = {
-  createUser,
+  createUser, //createUser: createUser,
 };
