+++ services/userPostService.js
@@ -1,0 +1,8 @@
+const userPostDao = require("../models/userPostDao");
+
+const userPost = async (id) => {
+  const listInfo = await userPostDao.userPost(id);
+  return listInfo;
+};
+
+module.exports = { userPost };
